requirements.txt
python-telegram-bot==20.7
requests==2.31.0
asyncio
aiohttp==3.9.1
APScheduler==3.10.4
python-dotenv==1.0.0